+++ test/Core.Test/AutoFixture/GroupFixtures.cs
@@ -1,37 +1,19 @@
-﻿using System;
-using AutoFixture;
-using Bit.Core.Models.Table;
-using Bit.Core.Test.AutoFixture.Attributes;
+﻿using Bit.Core.Test.AutoFixture.Attributes;
+using Bit.Core.Test.AutoFixture.OrganizationFixtures;
 
 namespace Bit.Core.Test.AutoFixture
 {
-    internal class GroupOrganization : ICustomization
-    {
-        public bool UseGroups { get; set; }
-
-        public void Customize(IFixture fixture)
-        {
-            var organizationId = Guid.NewGuid();
-
-            fixture.Customize<Organization>(composer => composer
-                .With(o => o.Id, organizationId)
-                .With(o => o.UseGroups, UseGroups));
-
-            fixture.Customize<Group>(composer => composer.With(g => g.OrganizationId, organizationId));
-        }
-    }
-
     internal class GroupOrganizationAutoDataAttribute : CustomAutoDataAttribute
     {
         public GroupOrganizationAutoDataAttribute() : base(
-            new SutProviderCustomization(), new GroupOrganization { UseGroups = true })
+            new SutProviderCustomization(), new OrganizationCustomization { UseGroups = true })
         { }
     }
 
     internal class GroupOrganizationNotUseGroupsAutoDataAttribute : CustomAutoDataAttribute
     {
         public GroupOrganizationNotUseGroupsAutoDataAttribute() : base(
-            new SutProviderCustomization(), new GroupOrganization { UseGroups = false })
+            new SutProviderCustomization(), new OrganizationCustomization { UseGroups = false })
         { }
     }
 }
